[Design] Change InfoCollectionCell Design(font, shadow)
