--- DailyPin/Utility/Constants.swift
@@ -1,42 +1,0 @@
-//
-//  Constants.swift
-//  DailyPin
-//
-//  Created by 김지연 on 2023/09/25.
-//
-
-import UIKit
-
-final class Constants {
-    
-    enum Color {
-        
-        static let background = UIColor(named: "backGroundColor")
-        static let basicText = UIColor(named: "textColor")
-        static let mainColor = UIColor(named: "mainColor")
-        static let subColor = UIColor(named: "SubColor")
-        static let pinColor = UIColor(named: "PinColor")
-        static let selectPinColor = UIColor(named: "SelectColor")
-        static let eventColor = UIColor(named: "EventColor")
-        static let cellBackgroundColor = UIColor(named: "CellBackgroundColor")
-        static let infoCellColor = UIColor(named: "InfoCell")
-        static let subBackgroundColor = UIColor(named: "SubBackgroundColor")
-        static let tintColor = UIColor.black
-        static let subTextColor = UIColor.lightGray
-        static let borderColor = UIColor.lightGray
-        static let placeholderColor = UIColor.lightGray
-    }
-    
-    enum Design {
-        static let cornerRadius: CGFloat = 10
-        static let plainFont: String = "NanumGothic"
-        static let boldFont: String = "NanumGothicBold"
-    }
-    
-    
-    
-    
-}
-
-
-
