+++ DailyPin/Utility/Constants/Color.swift
@@ -1,0 +1,28 @@
+//
+//  Color.swift
+//  DailyPin
+//
+//  Created by 김지연 on 3/5/24.
+//
+
+import UIKit
+
+extension Constants {
+    enum Color {
+        
+        static let background = UIColor(named: "backGroundColor")
+        static let basicText = UIColor(named: "textColor")
+        static let mainColor = UIColor(named: "mainColor")
+        static let subColor = UIColor(named: "SubColor")
+        static let pinColor = UIColor(named: "PinColor")
+        static let selectPinColor = UIColor(named: "SelectColor")
+        static let eventColor = UIColor(named: "EventColor")
+        static let cellBackgroundColor = UIColor(named: "CellBackgroundColor")
+        static let infoCellColor = UIColor(named: "InfoCell")
+        static let subBGColor = UIColor(named: "subBGColor")
+        static let tintColor = UIColor.black
+        static let subTextColor = UIColor.lightGray
+        static let borderColor = UIColor.lightGray
+        static let placeholderColor = UIColor.lightGray
+    }
+}
