+++ DailyPin/Utility/Constants/Constants.swift
@@ -1,0 +1,15 @@
+//
+//  Constants.swift
+//  DailyPin
+//
+//  Created by 김지연 on 2023/09/25.
+//
+
+import UIKit
+
+final class Constants {
+    
+}
+
+
+
