+++ DailyPin/Utility/Constants/Image.swift
@@ -1,0 +1,45 @@
+//
+//  Image.swift
+//  DailyPin
+//
+//  Created by 김지연 on 2023/10/08.
+//
+
+import Foundation
+import UIKit
+
+extension Constants {
+    
+    enum Image {
+        static let star = UIImage(systemName: "star.circle.fill")
+        static let mappin = UIImage(systemName: "mappin.circle")
+        static let mappinFill = UIImage(systemName: "mappin.circle.fill")
+        static let networkError = UIImage(systemName: "wifi.exclamationmark")
+        static let menuButton = UIImage(systemName: "ellipsis")
+        
+        static let calendar = UIImage(systemName: "calendar")
+        static let plus = UIImage(systemName: "plus")
+        static let noData = UIImage(systemName: "person.fill.questionmark")
+        static let noPlaceList = UIImage(systemName: "highlighter")
+        static let warning = UIImage(systemName: "exclamationmark.bubble")
+        static let placeList = UIImage(systemName: "list.bullet")
+        static let returnToday = UIImage(named: "todayCalendar")
+        
+        static let backButton = UIImage(systemName: "chevron.backward")
+        static let leftButton = UIImage(systemName: "chevron.left")
+        static let rightButton = UIImage(systemName: "chevron.right")
+        static let xmark = UIImage(systemName: "xmark")
+        
+        
+        static let selectPin = UIImage(named: "baseline_place_white_24pt")
+        static let pushPin = UIImage(named: "PushPin")
+        static let starImage = UIImage(named: "star")
+        static let google = UIImage(named: "Google")
+        
+        static let curLocation = UIImage(named: "CurrentLocation")
+        
+        
+        
+    }
+    
+}
